Added tweaks to make sure SEO optimizations are working
