+++ assets/data/personal-data.js
@@ -1,0 +1,35 @@
+export const personData = {
+    "@context": "https://schema.org",
+    "@type": "Person",
+    "name": "Pratyum Jagannath",
+    "alternateName": "PJ",
+    "description": "Full Stack Developer & Web3 Enthusiast",
+    "image": "https://pratyumjagan.in/images/profile-photo.jpg", // Update with actual image
+    "jobTitle": "Full Stack Developer",
+    "url": "https://pratyumjagan.in",
+    "sameAs": [
+      "https://linkedin.com/in/pratyumjagannath", // Update with actual links
+      "https://github.com/pratyum",
+      "https://twitter.com/pratyumjagan"
+    ],
+    "knowsAbout": [
+      "React",
+      "Next.js",
+      "Web3",
+      "Python",
+      "Node.js",
+      "Django",
+      "JavaScript",
+      "Full Stack Development"
+    ],
+    "alumniOf": {
+      "@type": "CollegeOrUniversity",
+      "name": "Nanyang Technological University",
+      "location": "Singapore"
+    },
+    "hasOccupation": {
+      "@type": "Occupation",
+      "name": "Full Stack Developer",
+      "occupationCategory": "Software Engineering"
+    }
+};
