+++ components/StructuredData.jsx
@@ -1,0 +1,8 @@
+export default function StructuredData({ data }) {
+    return (
+      <script
+        type="application/ld+json"
+        dangerouslySetInnerHTML={{ __html: JSON.stringify(data) }}
+      />
+    );
+}
