+++ package.json
@@ -9,6 +9,7 @@
     "lint:fix": "next lint --fix"
   },
   "dependencies": {
+    "@vercel/analytics": "^1.5.0",
     "framer-motion": "^10.18.0",
     "next": "^14.2.28",
     "react": "^18.3.1",
